Remove JsonEqual
